% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernden2d_ss.R
\name{kernden2d_ss}
\alias{kernden2d_ss}
\title{kernel density estimation for 2d using R.}
\usage{
kernden2d_ss(x, y, hx, hy, kernel)
}
\arguments{
\item{x}{the sample data of random variable x, it should be a squence.(vector)}

\item{y}{the sample data of random variable y, it should be a squence.(vector)}

\item{hx}{the bandwiths of x (numeric)}

\item{hy}{the bandwiths of y (numeric)}

\item{kernel}{the kernel function (character)}
}
\value{
a list of x , y and f.hat, (x,y) is coordinate point , f.hat is the estimation of density corresponding to (x,y)
}
\description{
2d kernel density function estimation by spherical-symmetric kernel function.
}
\examples{
\dontrun{
data(faithful)
res <- kernden2d_ss(faithful$eruptions, faithful$waiting, 0.23, 33, kernel="gaussian")
persp(res$x, res$y, res$f.hat, theta = -30, phi = 30)
}
}
