% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwmC}
\alias{rwmC}
\title{A random walk Metropolis sampler using RCPP}
\usage{
rwmC(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of normal distribution for the increment}

\item{x0}{the origial position of the chain}

\item{N}{the number of generating times}
}
\value{
a random sample for the standard Laplace distribution \code{n}
}
\description{
A random walk Metropolis sampler for the standard Laplace distribution using RCPP
}
\examples{
\dontrun{
N <- 2000
x0 <- 20
sigma <- 2
rwmC(sigma, x0, N)
}
}
